+++ examples/embassy/set_pico_w_led/src/main.rs
@@ -14,7 +14,7 @@ use embassy_time::Duration;
 use panic_persist as _;
 use picoserve::{
     response::DebugValue,
-    routing::{get, parse_path_segment},
+    routing::{get, get_service, parse_path_segment},
 };
 use rand::Rng;
 use static_cell::make_static;
@@ -152,15 +152,17 @@ async fn main(spawner: embassy_executor::Spawner) {
         picoserve::Router::new()
             .route(
                 "/",
-                get(|| picoserve::response::File::html(include_str!("index.html"))),
+                get_service(picoserve::response::File::html(include_str!("index.html"))),
             )
             .route(
                 "/index.css",
-                get(|| picoserve::response::File::css(include_str!("index.css"))),
+                get_service(picoserve::response::File::css(include_str!("index.css"))),
             )
             .route(
                 "/index.js",
-                get(|| picoserve::response::File::javascript(include_str!("index.js"))),
+                get_service(picoserve::response::File::javascript(include_str!(
+                    "index.js"
+                ))),
             )
             .route(
                 ("/set_led", parse_path_segment()),