+++ examples/web_sockets/src/main.rs
@@ -1,6 +1,9 @@
 use std::time::Duration;
 
-use picoserve::{response::ws, routing::get};
+use picoserve::{
+    response::ws,
+    routing::{get, get_service},
+};
 
 struct WebsocketHandler {
     tx: std::rc::Rc<tokio::sync::broadcast::Sender<String>>,
@@ -79,10 +82,11 @@ async fn main() -> anyhow::Result<()> {
         picoserve::Router::new()
             .route(
                 "/",
-                get(|| picoserve::response::File::html(include_str!("index.html"))),
+                get_service(picoserve::response::File::html(include_str!("index.html"))),
             )
-            .nest_service("/static", {
-                const STATIC_FILES: picoserve::response::Directory =
+            .nest_service(
+                "/static",
+                const {
                     picoserve::response::Directory {
                         files: &[
                             (
@@ -94,18 +98,19 @@ async fn main() -> anyhow::Result<()> {
                                 picoserve::response::File::css(include_str!("index.js")),
                             ),
                         ],
-                        sub_directories: &[],
-                    };
-
-                STATIC_FILES
-            })
+                        ..picoserve::response::Directory::DEFAULT
+                    }
+                },
+            )
             .route(
                 "/index.css",
-                get(|| picoserve::response::File::css(include_str!("index.css"))),
+                get_service(picoserve::response::File::css(include_str!("index.css"))),
             )
             .route(
                 "/index.js",
-                get(|| picoserve::response::File::javascript(include_str!("index.js"))),
+                get_service(picoserve::response::File::javascript(include_str!(
+                    "index.js"
+                ))),
             )
             .route(
                 "/ws",